fix(jsi): add cstdarg include and store context pointers as uintptr_t for Windows LLP64

+++ cpp/jsi/JSIContext.h
@@ -1,19 +1,22 @@
 #pragma once
 #include "JSINativeHeaders.h"
 #include <functional>
+#include <cstdint>
 #include <mutex>
 #include <unordered_map>
 #include <vector>
 
 namespace rnllama_jsi {
+    // Pointers are stored as uintptr_t (not long) so they stay full-width on
+    // Windows LLP64, where `long` is 32-bit and would truncate a 64-bit ptr.
     template<typename T>
     class ContextManager {
     private:
-        std::unordered_map<int, long> contextMap;
+        std::unordered_map<int, uintptr_t> contextMap;
         std::mutex contextMutex;
 
     public:
-        void add(int contextId, long contextPtr) {
+        void add(int contextId, uintptr_t contextPtr) {
             std::lock_guard<std::mutex> lock(contextMutex);
             contextMap[contextId] = contextPtr;
         }
@@ -23,7 +26,7 @@ namespace rnllama_jsi {
             contextMap.erase(contextId);
         }
 
-        long get(int contextId) {
+        uintptr_t get(int contextId) {
             std::lock_guard<std::mutex> lock(contextMutex);
             auto it = contextMap.find(contextId);
             return (it != contextMap.end()) ? it->second : 0;
@@ -34,9 +37,9 @@ namespace rnllama_jsi {
             return contextMap.size();
         }
 
-        std::vector<std::pair<int, long>> snapshot() {
+        std::vector<std::pair<int, uintptr_t>> snapshot() {
             std::lock_guard<std::mutex> lock(contextMutex);
-            std::vector<std::pair<int, long>> items;
+            std::vector<std::pair<int, uintptr_t>> items;
             items.reserve(contextMap.size());
             for (const auto& entry : contextMap) {
                 items.push_back(entry);
@@ -44,11 +47,11 @@ namespace rnllama_jsi {
             return items;
         }
         
-        void clear(std::function<void(long)> deleter = nullptr) {
+        void clear(std::function<void(uintptr_t)> deleter = nullptr) {
             // Take a snapshot and clear the map first, then delete objects.
             // This ensures any concurrent lookups via get() will return 0 (not found)
             // rather than a dangling pointer while deletion is in progress.
-            std::vector<long> toDelete;
+            std::vector<uintptr_t> toDelete;
             {
                 std::lock_guard<std::mutex> lock(contextMutex);
                 if (deleter) {
@@ -61,7 +64,7 @@ namespace rnllama_jsi {
             }
             // Delete outside the lock to avoid potential deadlocks
             if (deleter) {
-                for (long ptr : toDelete) {
+                for (uintptr_t ptr : toDelete) {
                     deleter(ptr);
                 }
             }
