+++ cpp/jsi/RNLlamaJSI.h
@@ -9,7 +9,7 @@ using namespace facebook;
 
 namespace rnllama_jsi {
     // Context management functions
-    void addContext(int contextId, long contextPtr);
+    void addContext(int contextId, uintptr_t contextPtr);
     void removeContext(int contextId);
     void setContextLimit(int64_t limit);
 #if defined(__ANDROID__)
